--- SAE-Puissance-4/src/main/java/etu/iuto/Main.java
@@ -1,7 +1,0 @@
-package etu.iuto;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello, World!");
-    }
-}
