+++ SAE-Puissance-4/src/main/java/etu/iuto/ServeurTcp.java
@@ -1,11 +1,13 @@
 package etu.iuto;
 
+import java.util.List;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.PrintWriter;
 import java.net.ServerSocket;
 import java.net.Socket;
+import java.util.ArrayList;
 
 
 public class ServeurTcp {
@@ -13,26 +15,36 @@ public class ServeurTcp {
     public static void main(String[] args) throws IOException{
 
         try{
+            List<Thread> client = new ArrayList<Thread>();
             ServerSocket serverSoket = new ServerSocket(1111);//choix du port
-            System.out.println("serveur en attente");
 
-            Socket clientSocket = serverSoket.accept();
-            System.out.println("connection etablie avec le client");
+            while(true){
+                System.out.println("serveur en attente");
+                Socket clientSocket = serverSoket.accept();
 
 
-            BufferedReader reader = new BufferedReader( new InputStreamReader(clientSocket.getInputStream()) );
-            PrintWriter writer = new PrintWriter(clientSocket.getOutputStream(),true);
+                Thread t = new Thread(new ClientTcp(clientSocket));
+                t.start();
+                System.out.println("connection etablie avec le client");
 
-            System.out.println(reader.readLine());
-            writer.println("message bien reçu !");
 
-            writer.println("by");
-            
+                BufferedReader reader = new BufferedReader( new InputStreamReader(clientSocket.getInputStream()) );
+                PrintWriter writer = new PrintWriter(clientSocket.getOutputStream(),true);
+                String read ="";
+                
+                read =reader.readLine();
+                System.out.println(read);
+                writer.println("message bien reçu !");
+                writer.println("quit");
+                
+                if (read != "quit") {
+                    reader.close();  
+                    writer.close();
+                    clientSocket.close();
+                    serverSoket.close();
+                }
+            }
 
-            reader.close();  
-            writer.close();
-            clientSocket.close();
-            serverSoket.close();
         }
         catch (Exception e){
             System.err.println("[erreur]" + e);
